requirements.txt
SpeechRecognition
pyaudio
pyttsx3
google-generativeai
python-dotenv